functions/create-user/requirements.txt
bcrypt
pyotp
psycopg2-binary
qrcode
Pillow
